% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_time}
\alias{check_time}
\title{Check order of observations}
\usage{
check_time(x, time_col = NULL, term = NULL)
}
\arguments{
\item{x}{}

\item{time_col}{}

\item{term}{}
}
\description{
Check order of observations
}
\author{
Héctor Miranda-Cebrián, \email{hectorm94@gmail.com}
}
