% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_data.R
\name{clean_community}
\alias{clean_community}
\title{Clean community data to use in other functions}
\usage{
clean_community(
  x,
  input_format = "wide",
  community_col = "comm",
  time_col = "time",
  taxa_col = "species",
  abundance_col = "abundance",
  na_zero = TRUE,
  filter_transient = TRUE,
  empty_years = FALSE,
  threshold = 0.3
)
}
\arguments{
\item{x}{A data.frame. Community matrix with time in rows and taxa in columns.}

\item{input_format}{Character. Format of the data to clean. One of "wide" or "long".}

\item{community_col}{Character. Name of column with the community identifier. Default "comm".}

\item{time_col}{Character. Name of column with time variable. Default "time".}

\item{taxa_col}{Character. Name of column with taxa names. Default "species".}

\item{abundance_col}{Character. Name of column with abundance values. Default "abundance".}

\item{na_zero}{Boolean. Replace missing values (NAs) with zeros (0). Default TRUE.}

\item{filter_transient}{Boolean. Filter transient species Default TRUE.}

\item{empty_years}{Boolean. Remove empty years. Default FALSE.}

\item{threshold}{Numeric. Minimum proportion (between 0 and 1) of time points with valid data to consider a species as transient. Default 0.3.}
}
\value{
A data.frame with community data ready to use in other functions.
}
\description{
Clean community data to use in other functions
}
\examples{
require(detrending)

# Clean data in long format
data(example_data_long)
clean_community(x = example_data_long, 
                input_format = "long",
                community_col = "comm",
                time_col = "time",
                taxa_col = "species",
                abundance_col = "abundance")

# Clean data in wide format
data(example_data_wide)
clean_community(x = example_data_wide, 
                input_format = "wide",
                community_col = "comm",
                time_col = "time")

}
\author{
Héctor Miranda-Cebrián, \email{hectorm94@gmail.com}
}
