% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_data.R
\name{clean_community_long}
\alias{clean_community_long}
\title{Clean community data in long format}
\usage{
clean_community_long(
  x,
  community_col = "comm",
  time_col = "time",
  taxa_col = "species",
  abundance_col = "abundance"
)
}
\arguments{
\item{x}{A data.frame. Community matrix with time in rows and taxa in columns.}

\item{community_col}{Character. Name of column with the community identifier. Default "comm".}

\item{time_col}{Character. Name of column with time variable. Default "time".}

\item{taxa_col}{Character. Name of column with taxa names. Default "species".}

\item{abundance_col}{Character. Name of column with abundance values. Default "abundance".}
}
\value{
A data.frame with the community data in long format.
}
\description{
Clean community data in long format
}
\author{
Héctor Miranda-Cebrián, \email{hectorm94@gmail.com}
}
