% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_data.R
\name{clean_community_wide}
\alias{clean_community_wide}
\title{Clean community data in wide format}
\usage{
clean_community_wide(
  x,
  community_col = "comm",
  time_col = "time",
  na_zero = TRUE,
  filter_transient = TRUE,
  empty_years = FALSE,
  threshold = 0.3
)
}
\arguments{
\item{x}{A data.frame. Community matrix with time in rows and taxa in columns.}

\item{community_col}{Character. Name of column with the community identifier. Default "comm".}

\item{time_col}{Character. Name of column with time variable. Default "time".}

\item{na_zero}{Boolean. Replace missing values (NAs) with zeros (0). Default TRUE.}

\item{empty_years}{Boolean. Remove empty years. Default FALSE.}

\item{threshold}{Numeric. Minimum proportion (between 0 and 1) of time points with valid data to keep a species. Default 0.3.}

\item{filter}{Boolean. Filter data. Default TRUE.}
}
\value{
A data.frame with the community data in wide format.
}
\description{
Clean community data in wide format
}
\author{
Héctor Miranda-Cebrián, \email{hectorm94@gmail.com}
}
