% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_exploration.R
\name{comm_expl}
\alias{comm_expl}
\title{Explore community data}
\usage{
comm_expl(
  x,
  by_timestep = FALSE,
  total = c("average", "overall"),
  community_col = "comm",
  time_col = "time",
  trend = FALSE
)
}
\arguments{
\item{x}{A data.frame. A community matrix of species abundance with years as rows and species as columns.}

\item{by_timestep}{Boolean. Get community information at each time step. Default FALSE.}

\item{total}{Character. Wether to compute diversity indices from the average relative abundance of species across years (overall) or the average of annula diversity indices.}

\item{community_col}{Character. Name of column with the community identifier.}

\item{time_col}{Character. Name of column with time variable.}
}
\value{
A data.frame.
}
\description{
Explore community data
}
