% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trends.R
\name{community_trends}
\alias{community_trends}
\title{Estimate population trends in a community}
\usage{
community_trends(
  x,
  time_col = "time",
  community_col = "comm",
  method = "loglinear",
  plot = FALSE,
  title = NULL
)
}
\arguments{
\item{x}{A data.frame. A community matrix of species abundances with time in rows and taxa in columns. Optionally it can include community and time columns.}

\item{time_col}{Character. Name of the column with time variable. Optional with default "time".}

\item{community_col}{Character. Name of column with the community identifier.}

\item{method}{Character. Method to estimate the trends, one of "dennis", "loglinear" or "rda". Default "dennis".}

\item{plot}{Boolean. Plot species abundances and their estimated trends. Default FALSE.}

\item{title}{Character. Title for the plot. Default NULL.}
}
\value{
A data.frame with the trend (in the natural logarithm scale) for each species in the community along with its variance and 95\% confidence interval.
}
\description{
Estimate population trends in a community
}
\examples{
require(detrending)

# Simulate community data with trends
comm_df <- sim_mvcomm(trend_mean = 0.3)

# Estimate trend for each species and plot them
community_trends(comm_df$sim_data, method = "loglinear", plot = TRUE)
}
