% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coefficient_variation.R
\name{cv_com_term}
\alias{cv_com_term}
\title{CV of a community}
\usage{
cv_com_term(x, total = TRUE, weighted = FALSE, term = "var")
}
\arguments{
\item{x}{A data.frame. Community matrix with time in rows and taxa in columns.}

\item{weighted}{Boolean. Weight the CV of each population by its average relative abundance per species across years.}

\item{term}{Character. Term to estimate the variance. One of "var" (for standard variance and covariance), "two" or "three" for Hills' two or three term local quadrat variance and covariance. Default "var".}
}
\description{
CV of a community
}
