% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_stability.R
\name{cv_mv}
\alias{cv_mv}
\title{Multivariate coefficient of variation}
\usage{
cv_mv(x, d = "euclidean", term = "var")
}
\arguments{
\item{x}{A community abundance matrix.}

\item{d}{Character. Distance metric to use. One of 'euclidean' or 'chord'. Default 'euclidean'.}

\item{term}{Character. Term to estimate the variance. One of "var" (for standard variance and covariance) "two" for Hills' two term local quadrat variance and covariance. Default "var".}
}
\description{
Multivariate coefficient of variation
}
