% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coefficient_variation.R
\name{cv_term}
\alias{cv_term}
\title{Calculate the coefficient of variation (CV) of a variable}
\usage{
cv_term(x, term = "var")
}
\arguments{
\item{x}{A numeric vector}

\item{term}{Character. Term to estimate the variance. One of "var" (for standard variance and covariance), "two" or "three" for Hills' two or three term local quadrat variance and covariance. Default "var".}
}
\description{
Calculate the coefficient of variation (CV) of a variable
}
