% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synchrony.R
\name{eta_gross}
\alias{eta_gross}
\title{Eta synchrony index}
\usage{
eta_gross(x, term = "var", time_col = "time", weighted = FALSE)
}
\arguments{
\item{x}{A data.frame. A community matrix of species abundance with years as rows and species as columns.}

\item{term}{Character. Term to estimate the variance. One of "var" (for standard variance and covariance), "two" or "three" for Hills' two or three term local quadrat variance and covariance. Default "var".}

\item{time_col}{Character. Name of the column with time variable. Optional, by default assumes that rows are in chronological order.}

\item{weighted}{Boolean. Weight index by average species relative abundances. Default FALSE.}
}
\value{
A numeric value.
}
\description{
This function estimates Gross et al. (2014) synchrony index Eta and its weighted version by Blüthgen et al. (2016) using a detrended version of variances based on Hill's 2 and 3 terms local quadratic variance.
}
\details{
\itemize{
\item Gross \emph{et al.}'s (2014) \eqn{\eta}:
\deqn{\eta = \dfrac{ 1 }{ S }\sum_{i=1}^{S}{corr(x_{i},\sum_{j \neq i}^{S}{x_{j})}}}
\item Blüthgen \emph{et al.}'s (2016) weighted version of \eqn{\eta}, \eqn{\eta_{w}}:
\deqn{\eta_{w} = \sum_{i=1}^{S}{p_{i} corr(x_{i},\sum_{j \neq i}^{S}{x_{j})}}}
}
}
\examples{
require(detrending)

# Simulate community data
comm_df <- sim_mvcomm()
eta_gross(x = comm_df$sim_data)
}
\references{
\itemize{
\item Gross, K., Cardinale, B. J., Fox, J. W., Gonzalez, A., Loreau, M., Wayne Polley, H., ... & van Ruijven, J. (2014). Species richness and the temporal stability of biomass production: a new analysis of recent biodiversity experiments. The American Naturalist, 183(1), 1-12.
\item Blüthgen, N., Simons, N. K., Jung, K., Prati, D., Renner, S. C., Boch, S., ... & Gossner, M. M. (2016). Land use imperils plant and animal community stability through changes in asynchrony rather than diversity. Nature Communications, 7(1), 10697.
}
}
