% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.R
\name{geom_seq}
\alias{geom_seq}
\title{Helper function to generate a geometric distribution of relative abundances}
\usage{
geom_seq(max_rel_abu, n_sp)
}
\arguments{
\item{max_rel_abu}{a single numeric value that represents the realtive
abundance of the most abundant species in the community.}

\item{n_sp}{number of species in the community.}
}
\value{
A vector with the length of the number of species containing the
relative abundances of the species.
}
\description{
Helper function to generate a geometric distribution of relative abundances
}
