% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{jenfish}
\alias{jenfish}
\title{Jennings–Fischer formula estimates a combined value of plant cover assuming overlap between plants.}
\usage{
jenfish(x, perc = FALSE)
}
\arguments{
\item{x}{Numeric. A vector of cover values.}

\item{perc}{Boolean. If the cover values are expressed as percentages or proportions. Default FALSE.}
}
\description{
Jennings–Fischer formula estimates a combined value of plant cover assuming overlap between plants.
}
\references{
\itemize{
\item Jennings, M. D., Faber-Langendoen, D., Loucks, O. L., Peet, R. K., & Roberts, D. (2009). Standards for associations and alliances of the US National Vegetation Classification. Ecological Monographs, 79(2), 173-199.
\item Fischer, H. S. (2015). On the combination of species cover values from different vegetation layers. Applied Vegetation Science, 18(1), 169-170.
}
}
\author{
Héctor Miranda-Cebrián, \email{hectorm94@gmail.com}
}
