% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_data.R
\name{metacoms_data}
\alias{metacoms_data}
\title{Prepare metacommunity data}
\usage{
metacoms_data(
  x,
  community_col = "comm",
  time_col = "time",
  taxa_col = "species"
)
}
\arguments{
\item{x}{A data.frame. Community matrix with time in rows and taxa in columns.}

\item{community_col}{Character. Name of column with the community identifier. Default "comm".}

\item{time_col}{Character. Name of column with time variable. Default "time".}

\item{taxa_col}{Character. Name of column with taxa names. Default "species".}
}
\value{
A data.frame with the community data ready to use by the \code{metacomstab_term()} function.
}
\description{
Prepare metacommunity data
}
\author{
@author Héctor Miranda-Cebrián, \email{hectorm94@gmail.com}
}
