% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot_com}
\alias{plot_com}
\title{Plot a community time series}
\usage{
plot_com(x, total = FALSE, title = NULL)
}
\arguments{
\item{x}{A data.frame. A community matrix of abundances with time in rows and taxa in columns.}

\item{total}{Boolean. Also plot the total abundance of the community by timestep. Default FALSE.}

\item{title}{Character. Title for the plot. Default NULL.}
}
\value{
A plot.
}
\description{
Plot a community time series
}
\examples{
require(detrending)

# Simulate data
comm_data <- sim_mvcomm()
plot_com(comm_data$sim_data)

}
\author{
Héctor Miranda-Cebrián, \email{hectorm94@gmail.com}
}
