% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synchrony.R
\name{psi_segrestin}
\alias{psi_segrestin}
\title{Psi}
\usage{
psi_segrestin(x, term = "var", ...)
}
\arguments{
\item{x}{A data.frame. A community matrix of species abundance with years as rows and species as columns.}

\item{term}{Character. Term to estimate the variance. One of "var" (for standard variance and covariance), "two" or "three" for Hills' two or three term local quadrat variance and covariance. Default "var".}

\item{time_col}{Character. Name of the column with time variable. Optional, by default assumes that rows are in chronological order.}
}
\value{
A single numeric value
}
\description{
Estimate Psi
}
\details{
\eqn{\psi} is estimated as follows:

\deqn{\psi = \sqrt{ \phi } ^ \alpha = \left( \dfrac{ \sigma_{x_{T}} }{ \sum_{i=1}^{S}{\sigma_{x_{i}}} } \right) ^ \alpha }

\deqn{\phi = \dfrac{ \sigma_{x_{T}}^2 }{ (\sum_{i=1}^{S}{\sigma_{x_{i}}})^2 }}

Where \eqn{\alpha} is a scaling factor calculated as:

\deqn{\alpha = \dfrac{ log(1 / 2) }{ log( \sum_{i=1}^{S}{\sigma_{x_{i}}^2} / (\sum_{i=1}^{S}{\sigma_{x_{i}}})^2 ) } }

Where \eqn{\sigma_{x}} is the standard deviation of a vector of abundances \eqn{x}, \eqn{S} is the number of species in the community, \eqn{x_{i}} is the abundance of species \eqn{i} across time steps, \eqn{x_{T}} is the sum of species abundances for each time step.
}
\references{
\itemize{
\item Segrestin, J., Götzenberger, L., Valencia, E., de Bello, F., & Lepš, J. (2024). A unified framework for partitioning the drivers of stability of ecological communities. Global Ecology and Biogeography, 33(5), e13828.
}
}
\author{
Jules Segrestin, \email{jsegrestin@gmail.com}

Héctor Miranda-Cebrián, \email{hectorm94@gmail.com}
}
