% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.R
\name{response}
\alias{response}
\title{Helper function to simulate response of species abundances to environmental cues}
\usage{
response(state = TRUE, bimodal = FALSE, mean = 1, sd = 1, n_sp, comp = FALSE)
}
\arguments{
\item{state}{Boolean. Defines if the species have some kind of trend. Default TRUE.}

\item{bimodal}{Boolean. If TRUE, half of the species respond opposite to the other
half, if FALSE, all species respond in the same direction. How many species
respond in a given direction also depends on the settings of mean and sd. Default FALSE.}

\item{mean}{Numeric. The mean of the normal distribution from which each of the species
responses is drawn.}

\item{sd}{Numeric. The standard deviation around the mean of the normal distribution from
which each of the species responses is drawn.}

\item{n_sp}{Numeric. Number of species in the community.}

\item{comp}{Boolean. If TRUE, species will exhibit compensatory dynamics. This is
simulated by having species of similar abundance to respond in opposite
directions. This argument is therefore only meaning full when having set
bimodal = TRUE. Default FALSE.}
}
\value{
The output is a vector with the length of the number of species, each
value representing the response of a species to hypotehtical cue (i.e. in the
context of the simulation, either an environmental signal or a longterm
monotonic abundance trend).
}
\description{
\code{response()} simulates the strength and direction with which the change in
species abundance across time responds to a stochastic environmental cue, or,
a general directional trend in abundance. This is only a helper function that
is used internally in \code{sim_mvcomm()} and \code{syngenr()}.
}
