% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.R
\name{sim_mvcomm}
\alias{sim_mvcomm}
\title{Simulate community from a multivariate normal}
\usage{
sim_mvcomm(
  n_sp = 10,
  years = 25,
  tot_abu = 200 * n_sp,
  power = 1.6,
  bound_pos = TRUE,
  corr = 0.5,
  p = 0.8,
  switch_trend = FALSE,
  trend_mean = 1,
  trend_sd = 0.01,
  bimodal_trend = FALSE
)
}
\arguments{
\item{n_sp}{Numeric. Number of species in the community.}

\item{years}{Numeric. NUmber of years to simulate.}

\item{tot_abu}{Numeric. Total abundance of the community.}

\item{power}{Numeric. Exponent of the Taylor's Power Law to estimate variance from mean abundance.}

\item{bound_pos}{Boolean. Bound abundance values to be positive. Default TRUE.}

\item{corr}{Numeric. Correlation between populations.}

\item{p}{Numeric. Dispersion parameter for a Dirichlet distribution. Controls the evenness of the community with lower values indicating less dominance.}

\item{switch_trend}{Boolean. Apply a trend to the data. Default FALSE.}

\item{trend_mean}{Numeric. Mean of the trend.}

\item{trend_sd}{Numeric. Standard deviation of the trend.}

\item{bimodal_trend}{Boolean. If TRUE half of the species have negative trends and half positive. Default FALSE.}
}
\value{
A named list with three elements:
\itemize{
\item \code{sim_data}: A data.frame with the simulated data, species in columns and time steps in rows.
\item \code{true_trend}: A named vector with the true mean trends of each simulated species.
\item \code{params}: A named vector with the parameters used to simulate the data.
}
}
\description{
Simulate community from a multivariate normal
}
\author{
Lars Götzenberger, \email{jsegrestin@gmail.com}

Jan Lepš, \email{suspa@prf.jcu.cz}

Héctor Miranda-Cebrián, \email{hectorm94@gmail.com}
}
