% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tpl}
\alias{tpl}
\title{Estimate of Taylor's Power Law}
\usage{
tpl(x, y)
}
\arguments{
\item{x}{Numeric. A vector of variances.}

\item{y}{Numeric. A vector of means.}
}
\value{
A named vector with coefficients alpha and beta for the relation between variance and mean according to Taylor's Power Law.
}
\description{
Estimate of Taylor's Power Law
}
