% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trends.R
\name{trend_dennis}
\alias{trend_dennis}
\title{Estimate mean annual growth rate following Dennis et al. (2001).}
\usage{
trend_dennis(x, time = NULL)
}
\arguments{
\item{x}{Numeric. A vector of abundances.}

\item{time}{Numeric. A vector with the time steps corresponding to each value in x.}
}
\value{
A named list with the mean annual growth rate in the natural logarithm scale and its variance.
}
\description{
Estimate mean annual growth rate following Dennis et al. (2001).
}
\references{
\itemize{
\item Dennis, B., Munholland, P. L., & Scott, J. M. (1991). Estimation of Growth and Extinction Parameters for Endangered Species. Ecological Monographs, 61(2), 115–143.
}
}
\author{
Héctor Miranda-Cebrián, \email{hectorm94@gmail.com}
}
