% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trends.R
\name{trend_loglinear}
\alias{trend_loglinear}
\title{Estimate mean annual growth rate using linear regression}
\usage{
trend_loglinear(x, time = NULL)
}
\arguments{
\item{x}{Numeric. A vector of abundances.}

\item{time}{Numeric. A vector with the time steps corresponding to each value in x.}
}
\value{
A numeric value with the slope of a linear regression of the log transformed abundances.
}
\description{
Estimate mean annual growth rate using linear regression
}
\author{
Héctor Miranda-Cebrián, \email{hectorm94@gmail.com}
}
