% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trends.R
\name{trend_mv}
\alias{trend_mv}
\title{Perform Redundancy Analysis on species abundances with time as explanatory variable to assess community level trends}
\usage{
trend_mv(
  x,
  time_col = "time",
  community_col = "comm",
  scale = TRUE,
  perm = 999
)
}
\arguments{
\item{x}{A data.frame. A community matrix of species abundances with time in rows and taxa in columns. Optionally it can include community and time columns.}

\item{time_col}{Character. Name of the column with time variable. Optional with default "time".}

\item{community_col}{Character. Name of column with the community identifier.}

\item{scale}{Boolean. Scale abundances to have mean 0 and standard deviation 1. Default TRUE.}

\item{perm}{Numeric. Number of permutations for significance testing. Default 999.}
}
\value{
An object of class \code{mv_trend}, a named list with three elements:
\itemize{
\item \code{anova}: A data.frame with the \emph{F-value} of the Redundancy Analysis and its p-value based on the number of permutations chosen.
\item \code{rda}: An object of class \code{rda} with the scores of the RDA.
\item \code{time}: A vector with the time variable used in the RDA.
}
}
\description{
Perform Redundancy Analysis on species abundances with time as explanatory variable to assess community level trends
}
\details{
This function estimates temporal trends in abundance at the community level by conducting Redundanncy Analysis (RDA) on species abundances with time as an explanaroty variable.
}
\examples{
require(detrending)

# Simulate community data with trends
comm_df <- sim_mvcomm(trend_mean = 0.3, bimodal_trend = TRUE)

# Estimate multivariate community trend
trend_mv(comm_df$sim_data, time_col = "time")

}
\references{
\itemize{
\item Legendre, P. and Legendre, L. (2012) Numerical Ecology. 3rd English ed. Elsevier.
}
}
\author{
Jan Lepš, \email{suspa@prf.jcu.cz}

Aleš Lisner, \email{lisnea00@jcu.cz}

Héctor Miranda-Cebrián, \email{hectorm94@gmail.com}
}
