% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_stability.R
\name{var_mv}
\alias{var_mv}
\title{Multivariate variance of community composition}
\usage{
var_mv(x, method = c("euclidean", "chord"))
}
\arguments{
\item{x}{A community abundance matrix.}

\item{method}{Character. Community dissimilarity metric to use. One of "euclidean" or "chord". Default "euclidean".}
}
\value{
A numeric value. The multivariate variance of the community.
}
\description{
Multivariate variance of community composition
}
\details{
The multivariate variance of community composition is defined as the average square Euclidean distance between annual observations (\eqn{X_{i}}) and the average community composition (\eqn{\overline{X}}) following:
\deqn{var_{mv} = \dfrac{ \sum_{i=1}^{t}{ED(X_{i}, \overline{X})^2} }{t-1}}
}
\examples{
require(detrending)

# Load and clean data
data(example_data_wide)
metacomm_df <- clean_community(x = example_data_wide,
                               input_format = "wide",
                               community_col = "comm",
                               time_col = "time",
                               filter_transient = FALSE)
comm_df <- metacomm_df[metacomm_df$comm == 1,][,-c(1:2)]

# Calculate multivariate variance
var_mv(x = comm_df, method = "euclidean") # 183.752
var_mv(x = comm_df, method = "chord") # 0.399
}
\author{
Jan Lepš, \email{suspa@prf.jcu.cz}

Aleš Lisner, \email{lisnea00@jcu.cz}

Héctor Miranda-Cebrián, \email{hectorm94@gmail.com}
}
