% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_stability.R
\name{var_t2mv}
\alias{var_t2mv}
\title{Multivariate two term local quadratic variance of community composition}
\usage{
var_t2mv(x, method = c("euclidean", "chord"))
}
\arguments{
\item{x}{A community abundance matrix.}

\item{method}{Character. Community dissimilarity metric to use. One of "euclidean" or "chord". Default "euclidean".}
}
\value{
A numeric value. The multivariate two term quadratic variance of the community.
}
\description{
Multivariate two term local quadratic variance of community composition
}
\details{
The multivariate two term local quadratic variance (\eqn{TTLQV_{mv2}}) of community composition is the detrended version of multivariate variance (\eqn{var_{mv}}):
\deqn{TTLQV_{mv2} = \dfrac{ \sum_{i=1}^{t-1}{ED(X_{i}, X_{i+1})^2} }{2(t-1)}}
Where \eqn{X_{i}} is the composition of the community at time \eqn{i} and \eqn{ED(X_{i}, X_{i+1})} the Euclidean distance between succesive time points.
}
\examples{
require(detrending)

# Load and clean data
data(example_data_wide)
metacomm_df <- clean_community(x = example_data_wide,
                               input_format = "wide",
                               community_col = "comm",
                               time_col = "time",
                               filter_transient = FALSE)
comm_df <- metacomm_df[metacomm_df$comm == 1,][,-c(1:2)]

# Calculate multivariate variance
var_t2mv(x = comm_df, method = "euclidean") # 193.893
var_t2mv(x = comm_df, method = "chord") # 0.370
}
\author{
Jan Lepš, \email{suspa@prf.jcu.cz}

Aleš Lisner, \email{lisnea00@jcu.cz}

Héctor Miranda-Cebrián, \email{hectorm94@gmail.com}
}
