% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_exploration.R
\name{comm_expl}
\alias{comm_expl}
\title{Explore community data}
\usage{
comm_expl(
  x,
  by_timestep = FALSE,
  total = "average",
  community_col = "comm",
  time_col = "time",
  trends = FALSE
)
}
\arguments{
\item{x}{A data.frame. A community matrix of species abundance with years as rows and species as columns.}

\item{by_timestep}{Boolean. Get community information at each time step. Default FALSE.}

\item{total}{Character. Wether to compute diversity indices from the average relative abundance of species across years (overall) or the average of annula diversity indices.}

\item{community_col}{Character. Name of column with the community identifier.}

\item{time_col}{Character. Name of column with time variable.}

\item{trends}{Boolean. Check for trends in species using linear regression on log-transformed abundances. Default FALSE.}
}
\value{
A named list:
\itemize{
\item \code{diversity}: A data.frame with several diversity metrics for each community.
\item \code{trends}: A data.frame with the estimated mean abundance trends of the species in each community.
}
}
\description{
Explore community data
}
