% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trends.R
\name{comm_trend}
\alias{comm_trend}
\title{Estimate population trends in a community}
\usage{
comm_trend(
  x,
  time_col = "time",
  method = "loglinear",
  offset = TRUE,
  plot = FALSE,
  title = NULL
)
}
\arguments{
\item{x}{A data.frame. A community matrix of species abundances with time in rows and taxa in columns. Optionally it can include community and time columns.}

\item{time_col}{Character. Name of the column with time variable. Optional with default "time".}

\item{method}{Character. Method to estimate the trends, one of "dennis", "loglinear" or "rda". Default "dennis".}

\item{offset}{Boolean. Add a small offset to each species (1\% of its mean abundance) to avoid problems with log(0). Default TRUE.}

\item{plot}{Boolean. Plot species abundances and their estimated trends. Default FALSE.}

\item{title}{Character. Title for the plot. Default NULL.}
}
\value{
A data.frame with the trend (in the natural logarithm scale) for each species in the community along with its variance and 95\% confidence interval.
}
\description{
Estimate population trends in a community
}
\examples{
require(detrending)

# Simulate community data with trends
comm_df <- sim_mvcomm(trend_mean = 0.3, bimodal_trend = TRUE)

# Estimate trend for each species and plot them
comm_trend(comm_df$sim_data, method = "loglinear", plot = TRUE)
}
