% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/composition_stability.R
\name{cv_mv}
\alias{cv_mv}
\title{Multivariate coefficient of variation}
\usage{
cv_mv(
  x,
  time_col = "time",
  method = c("euclidean", "chord"),
  term = c("var", "two")
)
}
\arguments{
\item{x}{A community abundance matrix.}

\item{time_col}{Character. Name of the column with time variable. Optional with default "time".}

\item{method}{Character. Community dissimilarity metric to use. One of "euclidean" or "chord". Default "euclidean".}

\item{term}{Character. Term to estimate the variance. One of "var" (for standard variance) "two" for Hill's two term local quadrat variance. Default "var".}
}
\value{
A numeric value. The multivariate coefficient of variation of the community estimated using the variance function of choice.
}
\description{
Multivariate coefficient of variation
}
\details{
The multivariate coefficient of variation is estimated by dividing the multivariate variance of community composition by the sample norm of the average composition.
}
\examples{
require(detrending)

# Load and clean data
data(example_data_wide)
metacomm_df <- clean_community(x = example_data_wide,
                               input_format = "wide",
                               community_col = "comm",
                               time_col = "time",
                               filter_transient = FALSE)
comm_df <- metacomm_df[metacomm_df$comm == 1,][,-c(1:2)]

# Calculate multivariate variance
cv_mv(x = comm_df, method = "euclidean", term = "var") # 0.958
cv_mv(x = comm_df, method = "euclidean", term = "two") # 0.961
}
\author{
Jan Lepš, \email{suspa@prf.jcu.cz}

Aleš Lisner, \email{lisnea00@jcu.cz}

Héctor Miranda-Cebrián, \email{hectorm94@gmail.com}
}
