% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synchrony.R
\name{eta_gross}
\alias{eta_gross}
\title{Eta synchrony index}
\usage{
eta_gross(x, term = "var", ...)
}
\arguments{
\item{x}{A data.frame. A community matrix of species abundance with years as rows and species as columns.}

\item{term}{Character. Term to estimate the variance. One of "var" (for standard variance and covariance), "two" or "three" for Hills' two or three term local quadrat variance and covariance. Default "var".}

\item{time_col}{Character. Name of the column with time variable. Optional, by default assumes that rows are in chronological order.}

\item{weighted}{Boolean. Weight the contribution of each species by its average abundance in the community. Default FALSE.}
}
\description{
This function estimates Gross et al. 2014 synchrony index Eta and its weighted version by Blüthgen et al. 2016 using a detrended version of variances based on Hill's 2 and 3 terms local quadratic variance.
}
\references{
\itemize{
\item Gross, K., Cardinale, B. J., Fox, J. W., Gonzalez, A., Loreau, M., Wayne Polley, H., ... & van Ruijven, J. (2014). Species richness and the temporal stability of biomass production: a new analysis of recent biodiversity experiments. The American Naturalist, 183(1), 1-12.
\item Blüthgen, N., Simons, N. K., Jung, K., Prati, D., Renner, S. C., Boch, S., ... & Gossner, M. M. (2016). Land use imperils plant and animal community stability through changes in asynchrony rather than diversity. Nature Communications, 7(1), 10697.
}
}
