% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synchrony.R
\name{phi_loreau}
\alias{phi_loreau}
\title{Phi synchrony index}
\usage{
phi_loreau(x, term = "var", ...)
}
\arguments{
\item{x}{A community matrix of species abundance with years as rows and species as columns.}

\item{term}{Character. Term to estimate the variance. One of "var" (for standard variance and covariance), "two" or "three" for Hills' two or three term local quadrat variance and covariance. Default "var".}

\item{time_col}{Character. Name of the column with time variable. Optional, by default assumes that rows are in chronological order.}
}
\value{
A single
}
\description{
This function estimates Loreau & Mazancourts 2008 Phi synchrony index using standard and detrended versions of variances based on Hill's 2 and 3 terms local quadratic variance.
}
\references{
\itemize{
\item Loreau, M., & de Mazancourt, C. (2008). Species synchrony and its drivers: neutral and nonneutral community dynamics in fluctuating environments. The American Naturalist, 172(2), E48-E66.
}
}
