% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot_com}
\alias{plot_com}
\title{Plot a community time series}
\usage{
plot_com(x)
}
\arguments{
\item{x}{A data.frame. A community matrix of abundances with time in rows and taxa in columns.}
}
\value{
A plot.
}
\description{
Plot a community time series
}
\author{
Héctor Miranda-Cebrián, \email{hectorm94@gmail.com}
}
