% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.R
\name{sim_mvcomm}
\alias{sim_mvcomm}
\title{Simulate community from a multivariate normal}
\usage{
sim_mvcomm(
  n_sp = 10,
  years = 25,
  tot_abu = 200 * n_sp,
  power = 1.8,
  corr = 0,
  even = 0.5,
  trend_mean = 0,
  trend_sd = 0.01
)
}
\arguments{
\item{n_sp}{Numeric. Number of species in the community.}

\item{years}{Numeric. NUmber of years to simulate.}

\item{tot_abu}{Numeric. Total abundance of the community.}

\item{power}{Numeric. Exponent of the Taylor's Power Law to estimate variance from mean abundance.}

\item{corr}{Numeric. Average correlation between populations.}

\item{even}{Numeric. The relative abundance (between 0 and 1) of the most abundant species. Controls the evenness of the community with lower values indicating more even communities. Alternatively, a vector of relative abundaces of length = n_sp.}

\item{trend_mean}{Numeric. Mean of the trend. Can be a single value (for a shared trend across species) or a vector of length = \code{n_sp} (for individual trends). Positive values indicate growth and negative ones, decline. Default 0 (no trend).}

\item{trend_sd}{Numeric. Standard deviation of the trend. Can be a single value or a vector of the same length as \code{trend_mean}.}
}
\value{
A named list with three elements:
\itemize{
\item \code{sim_data}: A data.frame with the simulated data including any possible trends. Species in columns and time steps in rows.
\item \code{baseline}: A data.frame with the simulated data without any trends. Species in columns and time steps in rows.
\item \code{true_trend}: A named vector with the true mean trends of each simulated species.
\item \code{params}: A named vector with the parameters used to simulate the data.
}
}
\description{
Simulate community from a multivariate normal
}
\examples{
require(detrending)

sim_mvcomm(n_sp = 15, years = 30)
}
\author{
Lars Götzenberger, \email{jsegrestin@gmail.com}

Jan Lepš, \email{suspa@prf.jcu.cz}

Héctor Miranda-Cebrián, \email{hectorm94@gmail.com}
}
